% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{autoplot.hrfals_fit}
\alias{autoplot.hrfals_fit}
\title{Autoplot method for hrfals_fit objects}
\usage{
\method{autoplot}{hrfals_fit}(x, vox = 1, ...)
}
\value{
A `ggplot` object.
}
\description{
Creates a ggplot showing the reconstructed HRF for a selected voxel.
}
