% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodbury_residualize.R
\name{configure_blas}
\alias{configure_blas}
\title{Configure BLAS threading}
\usage{
configure_blas(num_threads = parallel::detectCores())
}
\arguments{
\item{num_threads}{Number of threads to use. Defaults to all available cores.}
}
\value{
Invisibly, a character string describing the BLAS library.
}
\description{
Attempts to set the number of BLAS threads using the optional
\pkg{RhpcBLASctl} package. Returns the BLAS library name invisibly.
}
