% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.R
\name{construct_A_H_sparse}
\alias{construct_A_H_sparse}
\title{Construct Spatial h-update System Matrix}
\usage{
construct_A_H_sparse(lhs_block_list, lambda_s, L_mat, d, v)
}
\arguments{
\item{lhs_block_list}{list of length `v` with `d x d` matrices for each voxel.}

\item{lambda_s}{spatial regularization strength.}

\item{L_mat}{sparse Laplacian matrix (`v x v`).}

\item{d}{number of basis functions.}

\item{v}{number of voxels.}
}
\value{
`dgCMatrix` representing `A_H`.
}
\description{
Helper used by the spatial CF-ALS implementation. Forms the sparse
block matrix \eqn{A_H = BlockDiag(LHS_v) + \lambda_s (L \otimes I_d)}.
}
\keyword{internal}
