% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.R
\name{construct_preconditioner}
\alias{construct_preconditioner}
\title{Construct Block-Jacobi Preconditioner}
\usage{
construct_preconditioner(lhs_block_list, lambda_s, degree_vec, d, v)
}
\arguments{
\item{lhs_block_list}{list of per-voxel `d x d` matrices.}

\item{lambda_s}{spatial regularization strength.}

\item{degree_vec}{numeric degree vector from the Laplacian.}

\item{d}{number of basis functions.}

\item{v}{number of voxels.}
}
\value{
block-diagonal `dgCMatrix` preconditioner.
}
\description{
Forms a block-diagonal preconditioner matrix for the conjugate gradient solver.
}
\keyword{internal}
