% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hrf_cfals.R
\name{estimate_hrf_spatial_cfals}
\alias{estimate_hrf_spatial_cfals}
\title{Spatially-Regularised CF-ALS Convenience Wrapper}
\usage{
estimate_hrf_spatial_cfals(
  fmri_data_obj,
  fmrireg_event_model,
  target_event_term_name,
  hrf_basis_for_cfals,
  laplacian_obj,
  lambda_s_default = 0.1,
  ...
)
}
\arguments{
  \item{fmri_data_obj}{See \code{estimate_hrf_cfals}.}
  \item{fmrireg_event_model}{Design object describing the experiment.}
  \item{target_event_term_name}{Name of the event term to estimate.}
  \item{hrf_basis_for_cfals}{HRF basis object.}
  \item{laplacian_obj}{Object returned by \code{build_voxel_laplacian}.}
  \item{lambda_s_default}{Default spatial regularisation strength.}
  \item{...}{Additional arguments passed to \code{estimate_hrf_cfals}.}
}
\value{An \code{hrfals_fit} object.}
\description{Calls \code{estimate_hrf_cfals()} with a predefined non-zero spatial penalty.}
