% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{glance.hrfals_fit}
\alias{glance.hrfals_fit}
\title{Glance method for hrfals_fit objects}
\usage{
\method{glance}{hrfals_fit}(x, ...)
}
\arguments{
\item{x}{An `hrfals_fit` object.}

\item{...}{Unused.}
}
\value{
A data frame summarising the fit.
}
\description{
Returns a one-row summary with design information and mean goodness of fit.
}
