% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_interface.R
\name{hrfals}
\alias{hrfals}
\title{Fit HRFs using CF-ALS}
\usage{
hrfals(y, design, method = "cf_als", control = list(), ...)
}
\arguments{
  \item{y}{Numeric matrix of BOLD data (time points x voxels).}
  \item{design}{List produced by \code{create_cfals_design()} containing at least \code{event_model} and \code{hrf_basis}.}
  \item{method}{Estimation engine to use. Defaults to \code{"cf_als"}.}
  \item{control}{List of algorithm parameters overriding \code{hrfals_control_defaults()}.}
  \item{...}{Additional arguments passed to the estimation engine.}
}
\value{
An object of class \code{fmrireg_cfals_fit}.
}
\description{
High level interface for CF-ALS HRF estimation.
}
