% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_interface.R
\name{hrfals_control_defaults}
\alias{hrfals_control_defaults}
\title{Default control values for hrfals}
\usage{
hrfals_control_defaults()
}
\value{
Named list of default algorithm parameters used by \code{hrfals}.
}
\description{
Returns the default tuning parameters for \code{hrfals()}.
}
