% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_interface.R
\name{hrfals_from_design}
\alias{hrfals_from_design}
\title{Fit HRFs using CF-ALS (Design Interface)}
\usage{
hrfals_from_design(y, design, method = "cf_als", control = list(), ...)
}
\arguments{
\item{y}{Numeric matrix of BOLD data (time points \eqn{\times} voxels).}

\item{design}{List produced by [create_cfals_design()] containing at least
`event_model` and `hrf_basis`.}

\item{method}{Estimation engine to use. Passed to [estimate_hrf_cfals()].}

\item{control}{List of control parameters overriding defaults.}

\item{...}{Additional arguments passed to [estimate_hrf_cfals()].}
}
\value{
An object of class `hrfals_fit`.
}
\description{
Internal helper that dispatches to [estimate_hrf_cfals()] when a pre-built
design list is available. Control parameters are merged with
[hrfals_control_defaults()] via [modifyList()].
}
