% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lss_check_conditioning}
\alias{lss_check_conditioning}
\title{Diagnostic function to check matrix conditioning}
\usage{
lss_check_conditioning(A, lambda_ridge = 0)
}
\arguments{
\item{A}{Nuisance regressor matrix}

\item{lambda_ridge}{Ridge penalty}
}
\value{
List with condition number and other diagnostics
}
\description{
Diagnostic function to check matrix conditioning
}
