% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{plot.hrfals_fit}
\alias{plot.hrfals_fit}
\title{Plot method for hrfals_fit objects}
\usage{
\method{plot}{hrfals_fit}(x, vox = 1, ...)
}
\arguments{
\item{x}{An `hrfals_fit` object.}

\item{vox}{Voxel index to plot (default = 1).}

\item{...}{Additional graphical parameters passed to `plot()`.}
}
\value{
Invisibly returns the reconstructed HRF for the specified voxel.
}
\description{
Plot method for hrfals_fit objects
}
