% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfals_methods.R
\name{print.fmrireg_cfals_fit}
\alias{print.fmrireg_cfals_fit}
\title{Methods for fmrireg_cfals_fit Objects}
\usage{
\method{print}{fmrireg_cfals_fit}(x, ...)
}
\arguments{
\item{x, object}{An `fmrireg_cfals_fit` object.}

\item{...}{Additional arguments passed to underlying functions.}

\item{vox}{Index of the voxel to plot.}
}
\description{
Basic utilities for inspecting the results
of `fmrireg_hrf_cfals`.
}
