% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{print.hrfals_fit}
\alias{print.hrfals_fit}
\title{Print method for hrfals_fit objects}
\usage{
\method{print}{hrfals_fit}(x, ...)
}
\arguments{
\item{x}{An `hrfals_fit` object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for hrfals_fit objects
}
