% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{print.summary.hrfals_fit}
\alias{print.summary.hrfals_fit}
\title{Print method for summary.hrfals_fit objects}
\usage{
\method{print}{summary.hrfals_fit}(x, ...)
}
\arguments{
\item{x}{A `summary.hrfals_fit` object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for summary.hrfals_fit objects
}
