% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{residuals.hrfals_fit}
\alias{residuals.hrfals_fit}
\title{Residuals method for hrfals_fit objects}
\usage{
\method{residuals}{hrfals_fit}(object, ...)
}
\arguments{
\item{object}{An `hrfals_fit` object.}

\item{...}{Additional arguments (unused).}
}
\value{
Residual matrix stored in the fit object.
}
\description{
Residuals method for hrfals_fit objects
}
