% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{autoplot.hrfals_fit}
\alias{autoplot.hrfals_fit}
\title{Autoplot method for hrfals_fit objects}
\usage{
\method{autoplot}{hrfals_fit}(object, vox = 1, ...)
}
\arguments{
\item{object}{An `hrfals_fit` object.}

\item{vox}{Voxel index to plot (default = 1).}

\item{...}{Additional arguments (currently unused).}
}
\value{
A `ggplot` object.
}
\description{
Creates a ggplot showing the reconstructed HRF for a selected voxel.
}
