% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_interface.R
\name{hrfals_control_defaults}
\alias{hrfals_control_defaults}
\title{hrfals control defaults}
\usage{
hrfals_control_defaults()
}
\value{
Named list of defaults used by [hrfals()].
}
\description{
Returns a list of default tuning parameters for `hrfals()`.
}
