% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrfals_methods.R
\name{hrfals_fit}
\alias{hrfals_fit}
\title{Construct an `hrfals_fit` object}
\usage{
hrfals_fit(
  h_coeffs,
  beta_amps,
  method,
  lambdas,
  call,
  fmrireg_hrf_basis_used,
  target_event_term_name,
  phi_recon_matrix,
  design_info,
  residuals,
  bad_row_idx = integer(0),
  recon_hrf = NULL,
  gof = NULL
)
}
\arguments{
\item{h_coeffs}{Matrix of HRF basis coefficients (d x v).}

\item{beta_amps}{Matrix of condition amplitudes (k x v).}

\item{method}{Character string indicating the estimation method.}

\item{lambdas}{Numeric vector of regularisation parameters.}

\item{call}{The matched call to the wrapper function.}

\item{fmrireg_hrf_basis_used}{HRF basis object supplied to the wrapper.}

\item{design_info}{List with design metadata (d, k, n, v, fullXtX). When
predictor standardisation is used the vectors `predictor_means` and
`predictor_sds` are also included.}

\item{residuals}{Residual matrix from the projected data fit.}

\item{bad_row_idx}{Integer vector of time points that were zeroed due to NA
values.}

\item{recon_hrf}{Matrix of reconstructed HRF shapes.}

\item{gof}{Numeric vector of goodness-of-fit statistics per voxel.}
}
\value{
An object of class `hrfals_fit`.
}
\description{
Simple constructor used by higher-level functions to package
the results returned by the various CFALS engines.
}
